Better cross-platform assertions
